Move util up to global utils
